+++ www/go/core/acl/model/AclOwnerEntity.php
@@ -398,6 +398,17 @@ abstract class AclOwnerEntity extends AclEntity {
 		parent::check();
 	}
 
+
+	private static function checkEmptyAcls() {
+		foreach(self::find(['id', static::$aclColumnName])->where(static::$aclColumnName, '=', null) as $model) {
+			$model->createAcl();
+			if(!$model->save()) {
+				throw new SaveException($model);
+			}
+		}
+
+	}
+
 	/**
 	 * Executed when a database check is performed
 	 *
@@ -409,6 +420,9 @@ abstract class AclOwnerEntity extends AclEntity {
 	 * @throws Exception
 	 */
 	public static function checkAcls() {
+
+		self::checkEmptyAcls();
+
 		$table = static::getMapping()->getPrimaryTable();
 
 		//set owner and entity properties of acl