+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Files: New permission to show main screen or not. This was it's possible to use files only for other items like
   project, contacts etc.
 - Core: Database check fixes AclOwnerEntity without ACL set
+- Tickets: Refresh button in display panel
 
 23-03-2022 6.6.58
 - Projects: bugfix do not set default task list in tasklist combo